requirements.txt
python-dotenv
langchain
langgraph
langchain-mcp-adapters
langchain-openai
langgraph_supervisor
